pyproject.toml
[project]
name = "openai-agent-sdk-playground"
version = "0.1.0"
description = "OpenAI Agents SDK demos: joke runner and sales outreach with tools and handoffs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.29.0",
    "openai-agents>=0.12.5",
    "python-dotenv>=1.2.2",
    "sendgrid>=6.12.5",
]
